+++ src/main/java/at/porscheinformatik/seleniumcomponents/DefaultSeleniumEnvironment.java
@@ -53,7 +53,7 @@ public class DefaultSeleniumEnvironment implements SeleniumEnvironment
          *  When running on the server with browserstack it happens from time to time that browserstack is not reachable and a exception is thrown.
          *  Wait some time and try it again. Hopefully it works then.
          */
-        return SeleniumUtils
-            .keepTrying(SeleniumGlobals.getLongTimeoutInSeconds(), () -> driverFactory.createWebDriver(sessionName));
+        return SeleniumUtils.keepTrying(SeleniumGlobals.getLongTimeoutInSeconds(),
+            () -> driverFactory.createWebDriver(sessionName));
     }
 }